Average chromosome signal (coverage) dot plot

This function allows you to plot average chromosome signal for one or two selected samples. It takes as input the output of chr_coverage (16x2 R data frame).

chr_coverage_plot(coverageDataA, coverageDataB, genome, meanNorm = FALSE,
  yMax, onScreen = TRUE, fileName, colorA = "grey50", colorB = "green")

Arguments

coverageDataA
A 16x2 data frame of coverage: chromosome and average signal. No default.

coverageDataB
Optional 16x2 data frame of coverage: chromosome and average signal. No default.

genome
A string representing the genome used for mapping. No default.

meanNorm
Boolean indicating whether input data are normalized to genome-wide averages (meanNorm = TRUE). If so, a line is added at y = 1. Defaults to FALSE.

yMax
Optional number to be used as the max Y scale value in the plots. No default.

onScreen
Boolean indicating plots should be returned to the screen (onScreen = TRUE) or written to .pdf files (onScreen = FALSE). Defaults to TRUE.

fileName
A string to name the output .pdf file in case (onScreen = FALSE). No default.

colorA
Optional R color for sample A. Defaults to grey50.

colorB
Optional R color for sample B. Defaults to green.

Value

A dot plot of one or two samples, either on screen or as a .pdf file (in the working directory).

Examples

# NOT RUN {
chr_coverage_plot(WT, rec8, genome = 'SK1', onScreen = TRUE, colorB = 'red')

chr_coverage_plot(WT, dot1, genome = 'S288C', meanNorm = TRUE,
                  onScreen = FALSE, fileName='chr_coverage_WT_and_dot1.pdf')
# }