FACS density plot

This function allows you to plot a density plot across time points of FACS data. Written by Tovah Markowitz.

facs_density_plot(yeastLine, gates, type)

Arguments

yeastLine
A number present in the file names corresponding to the yeast strain number. No default.

gates
Optional gating limits as c(gateLower, gateUpper). Defaults to no gating.

type
A string representing the image file type, acepting either jpg or pdf. Defaults to jpg.

Value

A density plot of all time points with the same yeast strain, either on screen or as a file (in the working directory).

Details

The function is designed to search all .FCS files (named as "Cell YeastLine_TimePoint.fcs") for those with a particular yeast line and print a quick view of ungated FACS results to the screen. It also allows for the option of gating the data before writing the results to an image file in the working directory.

The function requires two R packages: "flowCore" and "flowViz". You can install the packages by running: source("http://bioconductor.org/biocLite.R") biocLite("flowCore") biocLite("flowViz")

(If the packages are missing you will see an error message with this information).

Note

If the number of the yeast line is found anywhere else in the name of any of the files in the working directory, this function WILL NOT work.

Examples

# NOT RUN {
facs_density_plot(119)

facs_density_plot(119, c(2000000, 7500000), 'pdf')
# }