Standard analysis of ChIP-seq experiments

This function will run the lab's standard analysis of ChIP-seq experiments, for which tab-separated wiggle data is generated. It will call different functions in the package to produce several .pdf files of analysis plots, written to a new folder in ".../LabShare/HTGenomics/Opening_act/". Note: When running on data aligned to the SK1 genome three of the plots are not produced: signal at sub-telomeric regions, signal at DSB hotspots and signal at axis binding sites. The reason for skipping the sub-telomeric signal analysis is the fact that the SK1 genome annotation contains inconsistencies at sub-telomeric regions. The reason for skipping the other two analysis is the fact that the reference data for DSB hotspots and Red1 binding sites were not available for SK1 at the time of writing this function. It is probably redundant to run this (long!) analysis for both genomes anyway, and using data mapped to the S288c reference genome should be preferred.

opening_act(wiggleData, relevantGenotype, chipTarget, sampleID,
  userInput = TRUE, runMetaORF = TRUE)

Arguments

wiggleData
As a list of the 16 chr wiggle data (output of readall_tab). No default.

relevantGenotype
String indicating the relevant strain mutations. Just use "WT", for example, if there are no relevant mutations. No default.

chipTarget
String indicating the ChIP target protein. No default.

sampleID
String indicating the sample ID, including the ID used in the analysis pipeline (with a date) and the read mapping conditions (see examples below). The function asks the user to check that the provided "sampleID" matches the required format before proceeding with the analysis. No default.

userInput
Boolean indicating whether to ask user to check the format of the sampleID argument. Defaults to TRUE.

runMetaORF
Boolean indicating whether to run the meta ORF analysis. This analysis typically takes about 30 minutes to run, so it may be useful to exclude it. Defaults to TRUE.

Value

A new folder in ".../LabShare/HTGenomics/Opening_act/" containing output plots (as .pdf files) of the following analysis:

Chromosome size bias

Signal at centromeres

Signal flanking rDNA

Signal at sub-telomeric regions (data mapped to S288c reference genome only)

Signal at DSB hotspots (data mapped to S288c reference genome only)

Signal at axis binding sites (data mapped to S288c reference genome only)

Signal at meta ORF

Examples

# NOT RUN {
opening_act(wiggleData=WT, relevantGenotype="WT", chipTarget="Red1",
            sampleID="AH119C-040114-sacCer3-2mis")
opening_act(set1_wiggle_data, "set1", "Red1", "AH8584b-16032016-sacCer3-2mis")
opening_act(rec8, "rec8", "Red1", "AH8115b-24042015-SacCer3-2mis",
            userInput=FALSE, runMetaORF=FALSE)
# }