Batch run of opening_act function

This function will take a CSV file containing the path to multiple wiggle data folders and argument labels and run opening_act function for each data set in the file. It allows you to run a batch of data sets autmatically.

opening_act_batch_run(inputDataFile, useReadr = TRUE, runMetaORF = TRUE,
  userInput = FALSE)

Arguments

inputDataFile
String indicating path to file containing instructions. The file must be in CSV format with the following columns (column names in the file must be reproduced exactly as below):

path Path to wiggle data.

relevantGenotype String indicating the relevant strain mutations, to be used as the argument to opening_act of the same name.

chipTarget String indicating the ChIP target protein, to be used as the argument to opening_act of the same name.

sampleID The sample ID, to be used as the argument to opening_act of the same name.

No default.

useReadr
Boolean indicating whether to use the much faster read_tsv() from Hadley Wickham's readr package instead of base R's read.table when reading the wiggle data. Used as the argument to readall_tab of the same name. Defaults to TRUE.

runMetaORF
Boolean indicating whether to run the meta ORF analysis. This analysis typically takes about 30 minutes to run, so it may be useful to exclude it. Used as the argument to opening_act of the same name. Defaults to TRUE.

userInput
Boolean indicating whether to ask user to check the format of the provided labels. Defaults to FALSE.

Value

The output of opening_act for each of the data sets included in the batch. Check the documentation of opening_act for more detail.

Examples

# NOT RUN {
opening_act_batch_run(inputDataFile="~/Desktop/inputs.csv")
opening_act_batch_run(inputDataFile="~/Desktop/inputs.csv", useReadr = FALSE,
                      runMetaORF = TRUE, userInput = FALSE)
# }