Compute mean signal and SD by position of output of signal_from_telomeres

Given the output of signal_from_telomeres, this function allows you to compute the mean ChIP-seq signal and standard deviation by position (as distance to telomere) either for all chromosomes together or in two separate groups for small (chrs 1, 3, and 6) and large chromosomes (all remaining chromosomes).

signal_from_telomeres_average(inputData, separateSmallAndLarge = FALSE)

Arguments

inputData
A list produced by calling signal_from_telomeres. No default.

separateSmallAndLarge
Boolean specifying whether to generate a single data frame for all chromosomes (separateSmallAndLarge = FALSE) or two data frames separating small (chrs 1, 3, and 6) and large chromosomes (all remaining chromosomes). Defaults to FALSE.

Value

If separateSmallAndLarge = FALSE, a dataframe with three columns:

distance_to_telomere Distance to telomere in bp

mean_signal Average ChIP-seq signal

sd Standard deviation

If separateSmallAndLarge = TRUE, a list of two dataframes (each with the same three columns explained above):

small_chrs Data frame of chromosomes 1, 3 and 6

large_chrs Data frame of remaining chromosomes

Examples

# NOT RUN {
signal_from_telomeres_average(WT_signal)

signal_from_telomeres_average(signalFromTelomeres, separateSmallAndLarge = TRUE)
# }