To make a scatter plot comparison of two wiggle files

This function is designed to work with function wiggle_scatter to allow for an overall generalized comparison of two data sets. Great for comparing replicates, checking controls, etc. The two functions are used as follows:

wiggle_scatter will extract the necessary information using wiggle_compress for all chromosomes

wiggle_scatter_plot will make a simple .eps file of the results.

Written by Tovah Markowitz

wiggle_scatter_plot(scatter, method = "spearman")

Arguments

scatter
Input should be output of wiggle_scatter.

method
Method used to calculate correlation value on plot (same as for cor function). Defaults to "spearman".

Value

Output is an .eps file with file name and axis names determined by columns of input data frame. The function is written so that the x-axis and the y-axis have the same range.

Examples

# NOT RUN {
a <- wiggle_scatter(red1, hop1, 4000)

wiggle_scatter_plot(a)
# }